requirements.txt
scipy
librosa
tensorflow-cpu
matplotlib
scikit-fuzzy
keras-visualizer